Got ca full log coming, still need to display it
