+++ Screens/Screens.Module.js
@@ -3,6 +3,7 @@ import NewEntryMenuScreen from "./NewEntryMenu/NewEntryMenuScreen";
 import FeedEntryScreen from "./FeedEntry/FeedEntryScreen";
 import NapEntryScreen from "./NapEntry/NapEntryScreen";
 import DiaperEntryScreen from "./DiaperEntry/DiaperEntryScreen";
+import LogScreen from "./Log/LogScreen";
 
 export {
   MainScreen,
@@ -10,4 +11,5 @@ export {
   FeedEntryScreen,
   NapEntryScreen,
   DiaperEntryScreen,
+  LogScreen,
 };
